// dashboard/src/utils/hasura-api/generated/schemas/nativeQueryRelationship.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { NativeQueryRelationshipUsingNativeQuery } from './nativeQueryRelationshipUsingNativeQuery';
import type { NativeQueryRelationshipUsingTable } from './nativeQueryRelationshipUsingTable';

export interface NativeQueryRelationship {
  name: string;
  /** @nullable */
  comment?: string | null;
  using:
    | NativeQueryRelationshipUsingNativeQuery
    | NativeQueryRelationshipUsingTable;
}